Use existing Docker image to get a working version of Python 2.6

+++ test.sh
@@ -1,9 +1,23 @@
 #!/bin/bash
 
+if [ -e "${HOME}/.bashrc" ]
+then
+  source "${HOME}/.bashrc"
+fi
+
 set -exo pipefail
 
-pip install pytest unittest2
+python --version
 
-python setup.py bdist_wheel
-pip install dist/future-*-none-any.whl
+if [ -e "/root/pip" ]
+then
+  pip install /root/pip/*.zip /root/pip/*.whl /root/pip/*tar.gz
+else
+  pip install pytest unittest2
+fi
+
+pytag="py${PYTHON_VERSION//./}"
+
+python setup.py bdist_wheel --python-tag="${pytag}"
+pip install dist/future-*-${pytag}-none-any.whl
 pytest tests/
